requirements.txt
streamlit>=1.31.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
svgwrite>=1.4.3
pandas>=2.0.0
